+++ src/pages/board/QuestionDetail.vue
